requirements.txt
streamlit
pandas
numpy
aiohttp
python-dotenv
